+++ src/screens/App/screens/Login/components/Login.jsx
@@ -1,8 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 import axios from "../../../../../shared/caller";
 import { useForm } from "../../../../../shared/Form/useForm";
-import InputField from "../../../../../shared/Auth/InputField.Auth";
+import InputField from "../../../../../shared/Auth/InputField";
 import { setUserPersistData } from "../../../../../shared/Auth/Login";
 import useAlert from "../../../../../hooks/useAlert";
