+++ src/screens/App/screens/SignUp/components/SignUp.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import axios from "../../../../../shared/caller";
-import InputField from "../../../../../shared/Auth/InputField.Auth";
+import InputField from "../../../../../shared/Auth/InputField";
 import { useForm } from "../../../../../shared/Form/useForm";
 import useAlert from "../../../../../hooks/useAlert";
