+++ src/shared/Components/product/ProductHalf.jsx
@@ -5,7 +5,7 @@ import {
   ProductPrice,
   ProductPurchase,
   ProductRating,
-} from "./Product.Components";
+} from "./ProductComponents";
 
 function ProductHalf() {
   return (
